Move tests and update README
