+++ functionaltests/tests/deep/8_18_test.go
@@ -24,16 +24,21 @@ import (
 	"github.com/elastic/apm-server/functionaltests/internal/steps"
 )
 
-func TestUpgrade_8_17_to_8_18_Snapshot(t *testing.T) {
+// Data streams get marked for lazy rollover by ES when something
+// changed in the underlying template(s), which in this case is
+// the apm-data plugin update for 8.19 and 9.1:
+// https://github.com/elastic/elasticsearch/pull/119995.
+
+func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
 	t.Parallel()
-	from := versionsCache.GetLatestSnapshot(t, "8.17")
-	to := versionsCache.GetLatestSnapshot(t, "8.18")
+	from := versionsCache.GetLatestSnapshot(t, "8.18")
+	to := versionsCache.GetLatestSnapshot(t, "8.19")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
 	}
 
-	scenarios := deepUpgradeILMTestScenarios(
+	scenarios := deepUpgradeLazyRolloverILMTestScenarios(
 		from.Version,
 		to.Version,
 		steps.APMErrorLogs{
@@ -51,10 +56,37 @@ func TestUpgrade_8_17_to_8_18_Snapshot(t *testing.T) {
 	}
 }
 
-func TestUpgrade_8_17_to_8_18_BC(t *testing.T) {
+func TestUpgrade_8_18_to_8_19_BC(t *testing.T) {
 	t.Parallel()
-	from := versionsCache.GetLatestVersionOrSkip(t, "8.17")
-	to := versionsCache.GetLatestBCOrSkip(t, "8.18")
+	from := versionsCache.GetLatestVersionOrSkip(t, "8.18")
+	to := versionsCache.GetLatestBCOrSkip(t, "8.19")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	scenarios := deepUpgradeLazyRolloverILMTestScenarios(
+		from.Version,
+		to.Version,
+		steps.APMErrorLogs{
+			functionaltests.TLSHandshakeError,
+			functionaltests.ESReturnedUnknown503,
+			functionaltests.RefreshCache503,
+			functionaltests.PopulateSourcemapFetcher403,
+		},
+	)
+	for _, scenario := range scenarios {
+		t.Run(scenario.Name, func(t *testing.T) {
+			t.Parallel()
+			scenario.Runner.Run(t)
+		})
+	}
+}
+
+func TestUpgrade_8_18_to_9_0_Snapshot(t *testing.T) {
+	t.Parallel()
+	from := versionsCache.GetLatestSnapshot(t, "8.18")
+	to := versionsCache.GetLatestSnapshot(t, "9.0")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
@@ -67,6 +99,35 @@ func TestUpgrade_8_17_to_8_18_BC(t *testing.T) {
 			functionaltests.TLSHandshakeError,
 			functionaltests.ESReturnedUnknown503,
 			functionaltests.RefreshCache503,
+			// TODO: remove once fixed
+			functionaltests.PopulateSourcemapFetcher403,
+		},
+	)
+	for _, scenario := range scenarios {
+		t.Run(scenario.Name, func(t *testing.T) {
+			t.Parallel()
+			scenario.Runner.Run(t)
+		})
+	}
+}
+
+func TestUpgrade_8_18_to_9_0_BC(t *testing.T) {
+	t.Parallel()
+	from := versionsCache.GetLatestVersionOrSkip(t, "8.18")
+	to := versionsCache.GetLatestBCOrSkip(t, "9.0")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	scenarios := deepUpgradeILMTestScenarios(
+		from.Version,
+		to.Version,
+		steps.APMErrorLogs{
+			functionaltests.TLSHandshakeError,
+			functionaltests.ESReturnedUnknown503,
+			functionaltests.RefreshCache503,
+			// TODO: remove once fixed
 			functionaltests.PopulateSourcemapFetcher403,
 		},
 	)
